adding psr-4 autoloading support
