+++ core/helper/PrepareMessage.php
@@ -4,7 +4,7 @@
  * Date: 18/12/13
  */
 
-namespace phpList;
+namespace phpList\helper;
 
 
 class PrepareMessage
